requirements.txt
PyMuPDF>=1.23.0
openpyxl>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
modal>=1.2.6
pandas>=2.0.0
json-repair>=0.30.0
